moving methods to helper files
